+++ packages/orchestrator-ui-components/src/components/WfoTable/WfoTable/WfoTableHeaderCell/WfoSortDirectionIcon.tsx
@@ -1,17 +1,29 @@
 import React, { FC } from 'react';
 
 import { useOrchestratorTheme } from '@/hooks';
-import { WfoArrowNarrowDown, WfoArrowNarrowUp } from '@/icons';
+import { WfoArrowNarrowDown, WfoArrowNarrowUp, WfoArrowsUpDown } from '@/icons';
 import { SortOrder } from '@/types';
 
+import { SORTABLE_ICON_CLASS } from './styles';
+
 export type WfoSortDirectionIconProps = {
-    sortDirection: SortOrder;
+    sortDirection?: SortOrder;
 };
 
 export const WfoSortDirectionIcon: FC<WfoSortDirectionIconProps> = ({
     sortDirection,
 }) => {
     const { theme } = useOrchestratorTheme();
+
+    if (!sortDirection) {
+        return (
+            <WfoArrowsUpDown
+                className={SORTABLE_ICON_CLASS}
+                css={{ visibility: 'hidden' }}
+                color={theme.colors.subduedText}
+            />
+        );
+    }
 
     return sortDirection === SortOrder.ASC ? (
         <WfoArrowNarrowUp