+++ packages/orchestrator-ui-components/src/icons/WfoMinusCircleOutline.tsx
@@ -14,7 +14,6 @@ export const WfoMinusCircleOutline: FC<WfoIconProps> = ({
         version="1.1"
         xmlns="http://www.w3.org/2000/svg"
     >
-        <title>icon/minus-circle-outline</title>
         <g
             id="Symbols"
             stroke="none"