+++ packages/orchestrator-ui-components/src/pages/metadata/WfoResourceTypesPage.tsx
@@ -91,6 +91,7 @@ export const WfoResourceTypesPage = () => {
             width: '90',
             renderData: (value) => <WfoFirstPartUUID UUID={value} />,
             renderDetails: (value) => value,
+            renderTooltip: (value) => value,
         },
         resourceType: {
             columnType: ColumnType.DATA,
@@ -133,6 +134,11 @@ export const WfoResourceTypesPage = () => {
                     ))}
                 </EuiBadgeGroup>
             ),
+            renderTooltip: (productBlocks) => {
+                return productBlocks.map((productBlock) => (
+                    <p key={productBlock.name}>- {productBlock.name}</p>
+                ));
+            },
         },
     };