+++ packages/orchestrator-ui-components/src/pages/metadata/WfoTasksPage.tsx
@@ -148,6 +148,14 @@ export const WfoTasksPage = () => {
                         ))}
                 </EuiBadgeGroup>
             ),
+            renderTooltip: (productTags) => {
+                return productTags
+                    ?.filter(onlyUnique)
+                    .sort((tagA, tagB) => tagA.localeCompare(tagB))
+                    .map((productTag) => (
+                        <p key={productTag}>- {productTag}</p>
+                    ));
+            },
         },
         createdAt: {
             columnType: ColumnType.DATA,
@@ -156,6 +164,7 @@ export const WfoTasksPage = () => {
             renderData: (date) => <WfoDateTime dateOrIsoString={date} />,
             renderDetails: parseIsoString(parseDateToLocaleDateTimeString),
             clipboardText: parseIsoString(parseDateToLocaleDateTimeString),
+            renderTooltip: parseIsoString(parseDateToLocaleDateTimeString),
         },
     };