+++ packages/orchestrator-ui-components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orchestrator-ui/orchestrator-ui-components
 
+## 2.3.0
+
+### Minor Changes
+
+-   0efcf20: 315 - Making the sort icon always visible in cases of small table columns.
+-   1b23a42: 459 Implements single click sorting in the table. Marks WfoSortButtons as deprecated.
+
+## 2.2.0
+
+### Minor Changes
+
+-   e228d98: Sorts grouped table groups
+-   d24efe6: 1381 Render mouseover tooltip for the new tables
+
+### Patch Changes
+
+-   60021b1: Fixes checkboxes without labels
+
 ## 2.1.1
 
 ### Patch Changes